+++ CHANGES.rst
@@ -11,6 +11,7 @@ Improvements:
  * Use SwiftLint to enforce Swift style and conventions (PR #2300).
  * Fix SWIFT_VERSION configuration in post install hook of Podfile (PR #2302).
  * Authentication: support SSO by using the fallback URL (#2307).
+ * Authentication: .well-known support (#2117).
 
 Bug fix:
  * Reskin: status bar text is no more readable on iPad (#2276).